Shader Implementation. #3
Implemented A shader through with Tutorial series two, Which uses Defines FragmentShader, VertexShader and the use of File reading, writing and processing Code.

Added Comments and directions to parts of code, and Added Fragment and Vertex Correctly to the Program.

This version As placeholder Elements, such as triangleVerticies that are not necessary In the upcoming releases.

This version Requires that, File reading is changed and added to header File, as well as implementation of Object classes