SpaceShip Angles #14
The spaceship Now animates itself, when moves to the side, by slightly tilting the wings.

Use : A and D keys In order to Test them out

Shooting  As Spacebar Has been Implemented with Sphere to be added that can shoot.

Implemented Rotation Mechanic using Vec3  to be able to calculate tilting by, x,y,z.